move app/ante to ante
