+++ ante/setup_test.go
@@ -1,12 +1,12 @@
 package ante_test
 
 import (
+	"github.com/evmos/ethermint/ante"
 	"math/big"
 
 	storetypes "cosmossdk.io/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/evmos/ethermint/app/ante"
 	"github.com/evmos/ethermint/tests"
 	evmtypes "github.com/evmos/ethermint/x/evm/types"
 )
