+++ ante/signverify_test.go
@@ -1,11 +1,11 @@
 package ante_test
 
 import (
+	"github.com/evmos/ethermint/ante"
 	"math/big"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/evmos/ethermint/app/ante"
 	"github.com/evmos/ethermint/tests"
 	evmtypes "github.com/evmos/ethermint/x/evm/types"
 )
